#98: Move addVariant, removeVariant and changeMasterVariant builders to the ProductUpdateActionUtils.
